Fixed bug which caused connections through a socket override to fail.
When communicating through a socket overlay dummy addresses are needed which can be mapped to the underlying layer (e.g. a Steam user ID). Upon connecting RakNet communicates the address used to reach the recipient so it can be added to it's list of external addresses. Should two peers communicate using the same dummy override address for each other, they will send messages to the loopback, assuming the target address is their own external address. This can be fixed by never allowing a dummy override address to be added as an external address.