Fixed RakNet GUID generation.
A mistake in the generation of the random 64bit RakNet GUID causes only about half of the bits to be set on the Windows platform. On other platforms 'gettimeofday' is used to generate the GUID, which can lead to collisions. Also using the 4 last bits of a short sleep duration as a source of randomness isn't sufficiently random so the granularity has been increased a bit.