Fixed threading issue in CancelConnectionAttempt.
A mutex is locked when getting entries in requestedConnectionQueue, but these entries are then manipulated outside of the locked area. When calling CancelConnectionAttempt these entries can be removed while still being manipulated causing all kinds of corruptions and crashes. This is now fixed by queuing the canceling of connection attempts.